+++ src/Contracts/HasLikes.php
@@ -116,10 +116,10 @@ interface HasLikes
      * Fetch records sorted by likes count.
      *
      * @param \Illuminate\Database\Eloquent\Builder $query
-     * @param string $sortDirection
+     * @param string $direction
      * @return \Illuminate\Database\Eloquent\Builder
      */
-    public function scopeSortedByLikesCount(Builder $query, $sortDirection = 'desc');
+    public function scopeOrderByLikesCount(Builder $query, $direction = 'desc');
 
     /**
      * Add a like for model by the given user.