+++ src/Contracts/LikeableService.php
@@ -133,10 +133,10 @@ interface LikeableService
      * Fetch records sorted by likes count.
      *
      * @param \Illuminate\Database\Eloquent\Builder $query
-     * @param string $sortDirection
+     * @param string $direction
      * @return \Illuminate\Database\Eloquent\Builder
      */
-    public function scopeSortedByLikesCount(Builder $query, $sortDirection = 'desc');
+    public function scopeOrderByLikesCount(Builder $query, $direction = 'desc');
 
     /**
      * Fetch likes counters data.